目录调整

+++ chapter_04/notify/app/server/internal/biz/tag.go
@@ -14,7 +14,7 @@ type Tag struct {
 	Uuid      string
 	Name      string
 	Desc      string
-	Status    int8
+	Status    int
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
@@ -22,7 +22,7 @@ type TemplateTagRelation struct {
 	Id           int
 	TemplateUuid string
 	TagUuid      string
-	Status       int8
+	Status       int
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
 }
@@ -31,7 +31,7 @@ type UserTagRelation struct {
 	Id        int
 	UserUuid  string
 	TagUuid   string
-	Status    int8
+	Status    int
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
